♻️ Refactor: 캘린더 페이지 레이아웃 수정 및 서버 로직 보충
