+++ client/src/components/common/MenuModal.tsx
@@ -48,9 +48,15 @@ const MenuModal = (): JSX.Element => {
     >
       <div className={classes.menu__con}>
         <div className={classes.menu__link_con}>
-          <Link to="/calendar">달력</Link>
-          <Link to="/my-page">마이페이지</Link>
-          <Link to="/diary-list">다이어리</Link>
+          <Link to="/calendar" onClick={() => dispatch(toggleMenuModal())}>
+            달력
+          </Link>
+          <Link to="/my-page" onClick={() => dispatch(toggleMenuModal())}>
+            마이페이지
+          </Link>
+          <Link to="/diary-list" onClick={() => dispatch(toggleMenuModal())}>
+            다이어리
+          </Link>
           <div>로그아웃</div>
         </div>
         <div className={classes.menu__bottom}>Made by. Team Maltesers 🐶</div>
