+++ client/src/styles/calendar/Dday.module.css
@@ -1,12 +1,13 @@
 /* Dday 컴포넌트 스타일 */
 .dday__con {
-  width: 301px;
-  height: 36.3125rem;
+  width: 18.8125rem;
+  height: 33.125rem;
   border: 2px solid #c7d6eb;
   display: flex;
   flex-direction: column;
   align-items: center;
   padding: 0.9375rem 1.1875rem;
+  margin-top: 0.625rem;
 }
 
 .dday__title {
