+++ client/src/styles/calendar/WeeklyCalendar.module.css
@@ -53,11 +53,15 @@
 /* WeeklyBody 컴포넌트 스타일 */
 .weekly__body_con {
   width: 92.8125rem;
-  height: 45.0625rem;
+  height: 43.75rem;
   border: 1px solid var(--sub-color);
   border-top: none;
   display: flex;
   overflow: scroll;
+}
+
+.weekly__body_con::-webkit-scrollbar {
+  display: none;
 }
 
 .weekly__body_side {
