--- server/data/databasejy.js
@@ -1,11 +1,0 @@
-const mysql = require("mysql2/promise");
-require("dotenv").config();
-
-const pool = mysql.createPool({
-  host: "localhost",
-  database: "daygrid",
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-});
-
-module.exports = pool;
